Adds an alternative way of representing interventions in event_raw
Event_raw lists the interventions associated with each one, regardless
of their type. It may be useful to subdivide them by type right away,
so it adds a further level of detail for the interventions.
This remains however commented on because the system does not currently
justify such a detailed breakdown

Signed-off-by: Davide Ferrari <davideferrari8@gmail.com>