Adds locking and unlocking APIs for Events
To request the lock of an event so that it can be modified consistently
and multiple access problems avoided, you can request a lock via a GET
request to /events/<int:id>/lock to which you must provide the user name
of the applicant in the managing_user parameter.
If this parameter is not provided, the lock is not granted.

In the same way you can request the unlock of an event. In this case,
only the user who requested the lock or an administrator can unlock the
event.

Signed-off-by: Davide Ferrari <davideferrari8@gmail.com>